It should be working now. Works for socket connetion for the screens

+++ CMS/src/main/java/com/github/cs_24_sw_3_09/CMS/services/DisplayDeviceService.java
@@ -27,7 +27,7 @@ public interface DisplayDeviceService {
 
     DisplayDeviceEntity addTimeSlot(Long id, Long timeslotId);
 
-    boolean connectScreen(Long deviceId);
+    boolean connectScreen(int deviceId);
 
-    boolean disconnectScreen(Long deviceId);
+    boolean disconnectScreen(int deviceId);
 }
